Create thin side navigation bar
