+++ locales/de-DE/App.json
@@ -157,7 +157,6 @@
   "Alephium": "Alephium",
   "WALLET": "BRIEFTASCHE",
   "CURRENT WALLET": "AKTUELLE BRIEFTASCHE",
-  "MENU": "MENÜ",
   "Overview": "Übersicht",
   "Lock": "Sperren",
   "Enter password": "Passwort eingeben",
