+++ locales/en-US/App.json
@@ -157,7 +157,6 @@
   "Alephium": "Alephium",
   "WALLET": "WALLET",
   "CURRENT WALLET": "CURRENT WALLET",
-  "MENU": "MENU",
   "Overview": "Overview",
   "Lock": "Lock",
   "Enter password": "Enter password",
