+++ locales/fr-FR/App.json
@@ -160,7 +160,6 @@
   "Alephium": "Alephium",
   "WALLET": "WALLET",
   "CURRENT WALLET": "WALLET ACTIF",
-  "MENU": "MENU",
   "Overview": "Vue d'ensemble",
   "Lock": "Verrouiller",
   "Enter password": "Entrez le mot de passe",
