+++ locales/vi-VN/App.json
@@ -157,7 +157,6 @@
   "Alephium": "Alephium",
   "WALLET": "VÍ",
   "CURRENT WALLET": "VÍ HIỆN TẠI",
-  "MENU": "MENU",
   "Overview": "Tổng quan",
   "Lock": "Khoá",
   "Enter password": "Nhập mật khẩu",
