+++ src/style/themes.ts
@@ -66,7 +66,7 @@ export const darkTheme: DefaultTheme = {
     accent: colord('#101012').alpha(0.32).toRgbString()
   },
   font: {
-    primary: 'rgba(255, 255, 255, 0.95)',
+    primary: '#e3e3e3',
     secondary: '#C0C0C0',
     tertiary: '#767679',
     contrastPrimary: 'rgba(0, 0, 0, 1)',
@@ -74,11 +74,11 @@ export const darkTheme: DefaultTheme = {
     highlight: '#ffb800'
   },
   border: {
-    primary: 'rgb(43, 43, 48)',
-    secondary: 'rgb(34, 34, 38)'
+    primary: 'rgba(255, 255, 255, 0.08)',
+    secondary: 'rgba(255, 255, 255, 0.04)'
   },
   shadow: {
-    primary: '0 2px 2px rgba(0, 0, 0, 0.25)',
+    primary: '0 4px 4px rgba(0, 0, 0, 0.25)',
     secondary: '0 10px 10px rgba(0, 0, 0, 0.3)',
     tertiary: '0 25px 25px rgba(0, 0, 0, 0.2)'
   },
